+++ tests/testthat/soc_read/soda.demo.socrata.com/resource/2asz-g9qq.json-05abc6.R
@@ -1,6 +1,6 @@
 structure(list(method = "GET", url = "https://soda.demo.socrata.com/resource/2asz-g9qq.json?%24select=photo_column&%24limit=10000", 
     status_code = 200L, headers = structure(list(Server = "nginx", 
-        Date = "Tue, 22 Jul 2025 20:42:15 GMT", `Content-Type` = "application/json;charset=utf-8", 
+        Date = "Mon, 18 Aug 2025 16:31:24 GMT", `Content-Type` = "application/json;charset=utf-8", 
         `Transfer-Encoding` = "chunked", Connection = "keep-alive", 
         `Access-Control-Allow-Origin` = "*", ETag = "\"Zm94dHJvdC4zMzQxNF82XzIyOU53LUZBcU4xZE5KMHNrUWUyZHBEWkJXS09EUQ-MQg5Jt1WXGNeP7tFkPCm_yxUPsw--gzip--gzip\"", 
         `X-SODA2-Fields` = "[\"photo_column\"]", `X-SODA2-Types` = "[\"photo\"]", 
@@ -9,8 +9,8 @@ structure(list(method = "GET", url = "https://soda.demo.socrata.com/resource/2as
         `Last-Modified` = "Mon, 16 Sep 2019 23:13:03 GMT", Vary = "Accept-Encoding", 
         `Content-Encoding` = "gzip", Age = "0", `X-Socrata-Region` = "aws-us-east-1-fedramp-prod", 
         `Strict-Transport-Security` = "max-age=31536000; includeSubDomains", 
-        `X-Socrata-RequestId` = "11e25ae4cbee80cc02c0b2534068c2cf"), class = "httr2_headers"), 
+        `X-Socrata-RequestId` = "1252a2cecf8163c06ccbe08510a46602"), class = "httr2_headers"), 
     body = charToRaw("[{\"photo_column\":\"ZXF6YilAOjRH26G2TH05mvHMTlDnLCwn6CysLQP18hc\"}]\n"), 
-    timing = c(redirect = 0, namelookup = 2e-05, connect = 0, 
-    pretransfer = 6.5e-05, starttransfer = 0.127005, total = 0.127038
+    timing = c(redirect = 0, namelookup = 1.9e-05, connect = 0, 
+    pretransfer = 0.000104, starttransfer = 0.224074, total = 0.224152
     ), cache = new.env(parent = emptyenv())), class = "httr2_response")