+++ NEWS.md
@@ -1,3 +1,9 @@
+# socratadata 0.1.1
+
+* V3 requests are supported with additional "soc_api_key_id" and 
+  "soc_api_key_secret" environment variables.
+* sf column is appropriately assigned when `alias = "replace"`.
+
 # socratadata 0.1.0
 
 * Initial CRAN submission.